Refactor: ApiService : 서비스 로직 추가 캡슐화 Client : 추가
