+++ investement/src/main/java/com/example/investment/news/service/client/ApiService.java
@@ -1,0 +1,28 @@
+package com.example.investment.news.service.client;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.http.HttpHeaders;
+import org.springframework.http.ResponseEntity;
+import org.springframework.stereotype.Service;
+
+@Service
+public class ApiService {
+
+    private final RestTemplateClient restTemplateClient;
+
+    @Value("${naver.api.clientId}")
+    private String clientId;
+
+    @Value("${naver.api.clientSecret}")
+    private String clientSecret;
+
+    public ApiService(RestTemplateClient restTemplateClient) {
+        this.restTemplateClient = restTemplateClient;
+    }
+
+    public ResponseEntity<String> get(String url, HttpHeaders headers) {
+        headers.set("X-Naver-Client-Id", clientId);
+        headers.set("X-Naver-Client-Secret", clientSecret);
+        return restTemplateClient.get(url, headers);
+    }
+}
