+++ investement/src/main/java/com/example/investment/news/service/client/RestTemplateClient.java
@@ -1,10 +1,12 @@
-package com.example.investment.news.common;
+package com.example.investment.news.service.client;
 
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpMethod;
 import org.springframework.http.ResponseEntity;
+
 import org.springframework.stereotype.Component;
+
 import org.springframework.web.client.RestTemplate;
 
 @Component
